backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
reportlab==4.0.9
Pillow==10.4.0
PyPDF2==3.0.1

